Déplacement de composants pour de moins gros fichiers
